requirements.txt
anyio==4.0.0
appdirs==1.4.4
attrs==23.1.0
beautifulsoup4==4.12.2   
black==23.11.0
bs4==0.0.1
certifi==2023.7.22       
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cssselect==1.2.0
fake-useragent==1.3.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
importlib-metadata==6.8.0
lxml==4.9.3
mypy-extensions==1.0.0
numpy==1.26.1
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.1
parse==1.19.1
pathspec==0.11.2
phantomjs==1.4.1
platformdirs==4.0.0
pycparser==2.21
pyee==8.2.2
pyppeteer==1.0.2
pyquery==2.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-telegram-bot==20.6
pytz==2023.3.post1
requests==2.31.0
requests-html==0.10.0
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tqdm==4.66.1
trio==0.23.1
trio-websocket==0.11.1
tzdata==2023.3
urllib3==1.26.18
w3lib==2.1.2
websockets==10.4
wsproto==1.2.0
zipp==3.17.0